feat(mookme): move command logic in dedicated runner classes
