+++ packages/mookme/src/commands/init.ts
@@ -1,24 +1,6 @@
 import commander from 'commander';
-import inquirer from 'inquirer';
-import fs from 'fs';
-
 import { GitToolkit } from '../utils/git';
-import {
-  addedBehaviorQuestion,
-  choiceQuestion,
-  confirmQuestion,
-  folderQuestion,
-  selectHookTypes,
-} from '../prompts/init';
-import logger from '../utils/logger';
-
-function createDirIfNeeded(path: string) {
-  if (!fs.existsSync(path)) {
-    fs.mkdirSync(path);
-  }
-}
-
-const clear = () => process.stdout.write('\x1Bc');
+import { InitOptions, InitRunner } from '../runner/init';
 
 export function addInit(program: commander.Command): void {
   program
@@ -29,160 +11,9 @@ export function addInit(program: commander.Command): void {
     .option('--packages-path <packages-path>', 'Provide packages path and skip the associated prompter')
     .option('--skip-types-selection', 'Skip hook types selection')
     .option('--yes', 'Skip confirmation prompter')
-    .action(async (opts) => {
+    .action(async (opts: InitOptions) => {
       const git = new GitToolkit();
-
-      if (opts.onlyHook) {
-        const hookTypes = await selectHookTypes(opts.skipTypesSelection);
-        git.writeGitHooksFiles(hookTypes);
-        process.exit(0);
-      }
-
-      let packagesPath: string;
-      if (typeof opts.packagesPath !== 'undefined') {
-        packagesPath = opts.packagesPath;
-      } else {
-        packagesPath = (await inquirer.prompt([folderQuestion('packagesPath')])).packagesPath;
-      }
-
-      let selectedPackages: string[] = [];
-      if (opts.packages) {
-        selectedPackages = opts.packages;
-      } else {
-        clear();
-        const moduleDirs = fs
-          .readdirSync(`./${packagesPath}`, { withFileTypes: true })
-          .filter((dirent) => dirent.isDirectory() && !dirent.name.startsWith('.'))
-          .map((dirent) => dirent.name);
-
-        clear();
-        const { packages } = (await inquirer.prompt([
-          choiceQuestion('packages', 'Select folders to hook :\n', moduleDirs),
-        ])) as { packages: string[] };
-        selectedPackages = selectedPackages.concat(packages);
-
-        let { addSubFolder } = (await inquirer.prompt([
-          confirmQuestion('addSubFolder', 'Do you wanna add a subfolder for packages ?'),
-        ])) as { addSubFolder: boolean };
-
-        while (addSubFolder) {
-          const { subPath } = (await inquirer.prompt([folderQuestion('subPath', packagesPath)])) as { subPath: string };
-          const fullSubFolderPath = packagesPath ? `./${packagesPath}/${subPath}` : `./${subPath}`;
-          const moduleDirs = fs
-            .readdirSync(`./${fullSubFolderPath}`, { withFileTypes: true })
-            .filter((dirent) => dirent.isDirectory() && !dirent.name.startsWith('.'))
-            .map((dirent) => dirent.name);
-
-          clear();
-          const { packages } = (await inquirer.prompt([
-            choiceQuestion('packages', 'Select folders to hook :\n', moduleDirs),
-          ])) as { packages: string[] };
-          selectedPackages = selectedPackages.concat(
-            packages.map((subPackage) => `${packagesPath ? `${packagesPath}/` : ''}${subPath}/${subPackage}`),
-          );
-
-          addSubFolder = (
-            (await inquirer.prompt([
-              confirmQuestion('addSubFolder', 'Do you wanna add a subfolder for packages ?'),
-            ])) as {
-              addSubFolder: boolean;
-            }
-          ).addSubFolder;
-        }
-      }
-
-      clear();
-      let addedBehavior;
-      if (opts.addedBehaviour) {
-        addedBehavior = opts.addedBehavior;
-      } else {
-        addedBehavior = (await inquirer.prompt([addedBehaviorQuestion])).addedBehavior;
-      }
-
-      const mookMeConfig = {
-        packagesPath: `.${packagesPath ? `/${packagesPath}` : ''}`,
-        packages: selectedPackages,
-        addedBehavior,
-      };
-
-      const packagesHooksDirPaths = selectedPackages.map((mod) => `${mookMeConfig.packagesPath}/${mod}/.hooks`);
-
-      const hookTypes = await selectHookTypes(opts.skipTypesSelection);
-
-      clear();
-      logger.info('\nThe following configuration will be written into .mookme.json:');
-      logger.log(JSON.stringify(mookMeConfig, null, 2));
-
-      logger.info('\n The follwowing git hooks will be created:');
-      hookTypes.forEach((t) => logger.log(`./.git/hooks/${t}`));
-
-      logger.info('\n`.hooks/*.local.env` will be appended into the following `.gitignore` files:');
-      selectedPackages
-        .map((mod) => `${mookMeConfig.packagesPath}/${mod}/.gitignore`)
-        .forEach((p) => logger.log(`- ${p}`));
-      logger.log(`- ./.gitignore`);
-
-      logger.info('\nThe following directories will also be created:');
-      packagesHooksDirPaths.forEach((hookDir) => logger.log(`- ${hookDir}`));
-      logger.log(`- ./.hooks`);
-
-      let confirm;
-
-      if (opts.yes) {
-        confirm = true;
-      } else {
-        confirm = (
-          await inquirer.prompt([
-            {
-              type: 'confirm',
-              name: 'confirm',
-              message: 'Do you confirm ?',
-              default: true,
-            },
-          ])
-        ).confirm;
-      }
-
-      if (confirm) {
-        logger.warning('\nWriting configuration...');
-        fs.writeFileSync('.mookme.json', JSON.stringify(mookMeConfig, null, 2));
-        logger.success('Done.');
-
-        logger.warning('Initializing hooks folders...');
-        createDirIfNeeded('./.hooks');
-        createDirIfNeeded('./.hooks/shared');
-        fs.writeFileSync('./.hooks/shared/.gitkeep', '');
-        createDirIfNeeded('./.hooks/partials');
-        fs.writeFileSync('./.hooks/partials/.gitkeep', '');
-        fs.writeFileSync(
-          './.hooks/pre-commit.json',
-          JSON.stringify(
-            {
-              steps: [
-                {
-                  name: 'Example hook',
-                  command: 'echo "Hello world!"',
-                },
-              ],
-            },
-            null,
-            2,
-          ),
-        );
-        logger.success('An example hook has been written in `./.hooks/pre-commit.json`');
-        packagesHooksDirPaths.forEach((hookDir) => {
-          createDirIfNeeded(hookDir);
-        });
-
-        git.writeGitHooksFiles(hookTypes);
-        const paths = mookMeConfig.packages.map((pkg) => `${mookMeConfig.packagesPath}/${pkg}`);
-        paths.push('./.gitignore');
-        git.writeGitIgnoreFiles(mookMeConfig.packages.map((pkg) => `${mookMeConfig.packagesPath}/${pkg}`));
-      }
-
-      logger.warning('\nAdding local hooks to .gitignore');
-      fs.appendFileSync(`./.gitignore`, `\n**/.hooks/*.local.json\n`, { flag: 'a+' });
-
-      logger.success('Your hooks are configured.');
+      const initRunner = new InitRunner(git);
+      await initRunner.run(opts);
     });
 }
